when a whitelist is configured disallow remote ips which are not in the whitelist to send a list of failing disks
on our server there are multiple VMs, one of them hosts the bot, so the host
connects via tcp to the bot and tells it when a disk is broken.
However this means any of the other VMs in the same virtual network could
send a list of broken disks too. The whitelist is there to prevent this.

TODO: check if a malicious VM could send packages with a fake source ip using a
tool such as scapy and the bot would still accept it as valid.